Fixed type issue causing incorrect schedules in EDF

+++ rclcpp/include/rclcpp/experimental/executors/events_executor/priority_events_queue.hpp
@@ -33,7 +33,7 @@ namespace executors
 
 struct PriorityEvent
 {
-  int priority;
+  int64_t priority;
   rclcpp::experimental::executors::ExecutorEvent event;
 };
 
@@ -75,7 +75,7 @@ public:
   void
   enqueue(const rclcpp::experimental::executors::ExecutorEvent & event) override
   {
-    int priority = priority_[event.entity_key];
+    int64_t priority = priority_[event.entity_key];
     rclcpp::experimental::executors::PriorityEvent single_event = {priority, event};
     single_event.event.num_events = 1;
     {
@@ -96,7 +96,7 @@ public:
   void
   enqueue_unsafe(const rclcpp::experimental::executors::ExecutorEvent & event)
   {
-    int priority = priority_[event.entity_key];
+    int64_t priority = priority_[event.entity_key];
     rclcpp::experimental::executors::PriorityEvent single_event = {priority, event};
     single_event.event.num_events = 1;
     event_queue_.push(single_event);
@@ -250,7 +250,7 @@ public:
   void
   enqueue(const rclcpp::experimental::executors::ExecutorEvent & event) override
   {
-    int priority;
+    int64_t priority;
     if (event.type == rclcpp::experimental::executors::ExecutorEventType::TIMER_EVENT) {
       priority = static_cast<const rclcpp::TimerBase *>(event.entity_key)->get_arrival_time();
     } else {
@@ -277,7 +277,7 @@ public:
   void
   enqueue_unsafe(const rclcpp::experimental::executors::ExecutorEvent & event)
   {
-    int priority;
+    int64_t priority;
     if (event.type == rclcpp::experimental::executors::ExecutorEventType::TIMER_EVENT) {
       priority = static_cast<const rclcpp::TimerBase *>(event.entity_key)->get_arrival_time();
     } else {
@@ -316,7 +316,7 @@ public:
   void
   enqueue(const rclcpp::experimental::executors::ExecutorEvent & event) override
   {
-    int priority;
+    int64_t priority;
     if (event.type == rclcpp::experimental::executors::ExecutorEventType::TIMER_EVENT) {
       priority = static_cast<const rclcpp::TimerBase *>(event.entity_key)->get_period();
     } else {
@@ -343,7 +343,7 @@ public:
   void
   enqueue_unsafe(const rclcpp::experimental::executors::ExecutorEvent & event)
   {
-    int priority;
+    int64_t priority;
     if (event.type == rclcpp::experimental::executors::ExecutorEventType::TIMER_EVENT) {
       priority = static_cast<const rclcpp::TimerBase *>(event.entity_key)->get_period();
     } else {
